add comments to ipa related code
