+++ src/InteractivePathAnalysis/ncriticalpathreportparser.h
@@ -15,6 +15,9 @@ struct Line {
     Role role;
 };
 
+/** 
+ * Helper structure for Path Information, containing parsed data for a specific path. 
+ */
 struct PathInfo {
     int index = -1;
     std::string start;
@@ -23,6 +26,10 @@ struct PathInfo {
     std::string id() const { return start + ":" + end;}
     bool isValid() const { return (index != -1); }
 };
+
+/** 
+ * Represents minimal block information transmitted into an Item. 
+ */
 struct Block {
     std::vector<Line> lines;
     bool isPath() { return pathInfo.isValid(); }
@@ -31,5 +38,14 @@ struct Block {
 };
 using BlockPtr = std::shared_ptr<Block>;
 
-std::vector<BlockPtr> parsePathData(const std::vector<std::string>& lines);
+/**
+ * @brief Parser for the Critical Path Report generated by VPR.
+ * 
+ * This parser is designed to process the Critical Path Report output generated by VPR (Versatile Place and Route) tool.
+ */
+class NCriticalPathReportParser {
+public:
+    static std::vector<BlockPtr> process(const std::vector<std::string>& lines);
+};
+
